Make the graphite reporter from puppet accept a prefix.
So we don't fill the root namespace in graphite, this updates the report
processor to accept config from the YAML (of course) and use that, plus
'.' to add more hierarchy to graphite.

Will break things that look in the current place, but our graphite is
mostly broken anyway at present.