Agent config changes
This commit modifies the agent configuration in two ways:

First, it removes pre and postrun ubuntu commands. These are
installed by default with the standard Ubuntu repository, but
not with the Puppetlabs repository. Since these options are
configured to be present and there's no way to remove them
through this module, it might be best if they are removed.

Second, $report_server and $ca_server are renamed to $real_* to
satisfy the default fallback value of $server.