(maint) Add default_method parameter
This adds a default method paramater for those systems that may not wish
to default to 'cron', such as Windows or OpenBSD.

This was a problem currently for OpenBSD, because the service check that
disables the service, kills the running agent, so the run is never
completed.

This was a problem for Windows, since there is no cron available.

Both use cases are handled by this code.