(optionally) configure environmentpath
Starting puppet 3.5.0, we can now use a single directive to configure a
common pattern of modules and manifests in different environments:
`environmentpath`.

Experience shows that this setting should be in the `[main]` section, or
else things start getting wonky.

When setting `environmentpath`, we have to neglect the `modulepath` and
`manifestpath` settings.

It is not be within the scope of this new directive to extend to
extend `config_version_cmd`.

An optional `environment.conf` is presumed to be provided by the user,
via git.

This pr addresses #93