Fix broken nginx-unicorn support addresses puppet-puppet issue 119 by switching from using the default location in the vhost to using an explicitly defined location, in order to avoid the default headers that cannot be unset on the default location.
Add beaker tests to test for this problem
Note that the beaker tests for this pass, but there are still failures.
I think it's better to get this fixed and add tests rather than blocking
on fixing everything.

Extend vagrantfile
Not specific to this issue, just came up as a convenience