(issue #164) fix metadata.json
This removes the dependency on ploperations/interval, because it's no
longer used, and resolves open-ended dependencies by restricting them to
the current major version. This also changes the license identifier so
it can be programatically identified by Puppet Forge.

Without this, we do not conform to the forge recommended style guide.