Change puppet_cmd path on Darwin
Puppet Labs does not provide a repository for Puppet on OSX, nor are the
Mac Ports or Homebrew official packages.  Given that Puppet Labs does,
however, ship packages on the downlaods site, we can manually install
those packages and use them.

https://docs.puppetlabs.com/guides/install_puppet/install_osx.html

This work updates the path for the Puppet command that cron uses to
execute agent runs.