Refactor the way packages are handled
This code allows the package for both the master and the agent to be
named the same, while avoiding namespace collisions using the union()
function.  An added benefit here is that the user may specify a custom
package name for the master or the agent.  This may come in handy for
users who wish to use something newer, like puppet-server.