Add OpenBSD server support
Without this change, the module adds incorrect values to puppet.conf for
the user and group that the puppetmasterd process should run as.  This
corrects that behavior and adds a test.

Here also we implement the flatten function so in cases where the
modulepath is not an array, the run is able to continue unimpeded.