Test for correct OpenBSD agent service management
Without this change, the module attempts to manage the service file
configuration on OpenBSD due to it being expected by default.  This is
not so on OpenBSD and causes the run to break.  This fix simply adds the
undef value for the option and includes a test to avoid future
breakages.