Move [main] config items to puppet::config class
This relocates the configuration items that belong in the [main] section
of the puppet.conf file to be namespaced under the puppet::config class.
This follows the pattern for the agent and the master to have specific
pieces belong to each section.  Without this, the puppet::agent::config
is a bit messy, since it places configuration items into the main
section which couples the puppet::agent class to the puppet::master
class.